Rename safetensor_utils to checkpoint_utils
